+++ pythogic/base/Symbol.py
@@ -56,3 +56,19 @@ class PredicateSymbol(Symbol):
 
     def _members(self):
         return (self.name, self.arity)
+
+
+class TrueSymbol(Symbol):
+    def __init__(self):
+        super().__init__("⊤")
+
+class FalseSymbol(Symbol):
+    def __init__(self):
+        super().__init__("⊥")
+
+class LastSymbol(Symbol):
+    def __init__(self):
+        super().__init__("Last")
+
+
+DUMMY_SYMBOL = Symbol("dummy_proposition")